requirements.txt

async-timeout==3.0.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.0
bulky==1.0.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
codetiming==1.2.0
colorama==0.4.3
decorator==4.4.2
deprecation==2.1.0
distlib==0.3.0
fake-useragent==0.1.11
filelock==3.0.12
Flask==1.1.2
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.8
frozendict==1.2
httplib2==0.17.3
idna==2.7
ipython==7.14.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jsonify==0.5
MarkupSafe==1.1.1
multidict==4.7.5
numpy==1.18.4
packaging==20.3
pandas==1.0.3
parso==0.7.0
pause==0.2
pickleshare==0.7.5
Pillow==7.1.2
prompt-toolkit==3.0.5
protobuf==3.11.3
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyEX==0.2.1
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.23.0
rsa==4.0
selenium==3.141.0
six==1.14.0
socketIO-client-nexus==0.7.6
soupsieve==2.0
sseclient==0.0.26
temporal-cache==0.1.1
traitlets==4.3.3
typeguard==2.7.1
uritemplate==3.0.1
urllib3==1.25.9
useragent==0.1.1
virtualenv==20.0.20
wcwidth==0.1.9
websocket-client==0.57.0
Werkzeug==1.0.1
yarl==1.4.2
pip~=20.1
wheel~=0.34.2
setuptools~=41.2.0
python-dotenv~=0.13.0